+++ pioreactor/actions/pump.py
@@ -153,6 +153,29 @@ def _get_calibration(pump_type: str) -> structs.AnyPumpCalibration:
             )
 
 
+def _publish_pump_action(
+    pump_action: str,
+    ml: pt.mL,
+    mqtt_client: Client,
+    unit: Optional[str] = None,
+    experiment: Optional[str] = None,
+    source_of_event: Optional[str] = None,
+) -> structs.DosingEvent:
+    dosing_event = structs.DosingEvent(
+        volume_change=ml,
+        event=pump_action,
+        source_of_event=source_of_event,
+        timestamp=current_utc_datetime(),
+    )
+
+    mqtt_client.publish(
+        f"pioreactor/{unit}/{experiment}/dosing_events",
+        encode(dosing_event),
+        qos=QOS.EXACTLY_ONCE,
+    )
+    return dosing_event
+
+
 def _pump_action(
     pump_type: str,
     unit: Optional[str] = None,
@@ -236,22 +259,15 @@ def _pump_action(
             assert isinstance(duration, pt.Seconds)
 
             # publish this first, as downstream jobs need to know about it.
-            dosing_event = structs.DosingEvent(
-                volume_change=ml,
-                event=action_name,
-                source_of_event=source_of_event,
-                timestamp=current_utc_datetime(),
-            )
-
-            client.publish(
-                f"pioreactor/{unit}/{experiment}/dosing_events",
-                encode(dosing_event),
-                qos=QOS.EXACTLY_ONCE,
+            dosing_event = _publish_pump_action(
+                action_name, ml, client, unit, experiment, source_of_event
             )
 
             pump_start_time = time.monotonic()
 
-            if not continuously:
+            if manually:
+                pass
+            elif not continuously:
                 pump.by_duration(duration, block=False)
 
                 # how does this work? What's up with the (or True)?
@@ -399,6 +415,7 @@ add_alt_media = partial(_pump_action, "alt_media")
 @click.option("--ml", type=float)
 @click.option("--duration", type=float)
 @click.option("--continuously", is_flag=True, help="continuously run until stopped.")
+@click.option("--manually", is_flag=True, help="The media is manually added (don't run pumps)")
 @click.option(
     "--source-of-event",
     default="CLI",
@@ -410,6 +427,7 @@ def click_add_alt_media(
     duration: Optional[pt.Seconds],
     continuously: bool,
     source_of_event: Optional[str],
+    manually: bool,
 ):
     """
     Remove waste/media from unit
@@ -424,6 +442,7 @@ def click_add_alt_media(
         source_of_event=source_of_event,
         unit=unit,
         experiment=experiment,
+        manually=manually,
     )
 
 
@@ -466,6 +485,7 @@ def click_remove_waste(
 @click.option("--ml", type=float)
 @click.option("--duration", type=float)
 @click.option("--continuously", is_flag=True, help="continuously run until stopped.")
+@click.option("--manually", is_flag=True, help="The media is manually added (don't run pumps)")
 @click.option(
     "--source-of-event",
     default="CLI",
@@ -477,6 +497,7 @@ def click_add_media(
     duration: Optional[pt.Seconds],
     continuously: bool,
     source_of_event: Optional[str],
+    manually: bool,
 ):
     """
     Add media to unit
@@ -491,4 +512,5 @@ def click_add_media(
         source_of_event=source_of_event,
         unit=unit,
         experiment=experiment,
+        manually=manually,
     )